Prevent manual deletion of published documents
As a good practice we should enforce our UI rules in the backend.

+++ app/controllers/documents_controller.rb
@@ -23,11 +23,13 @@ class DocumentsController < ApplicationController
 
   def confirm_delete_draft
     document = Document.find_by_param(params[:id])
+    raise "Trying to delete a live document" if document.has_live_version_on_govuk
     redirect_to document_path(document), confirmation: "documents/show/delete_draft"
   end
 
   def destroy
     document = Document.find_by_param(params[:id])
+    raise "Trying to delete a live document" if document.has_live_version_on_govuk
     DocumentPublishingService.new.discard_draft(document)
     document.destroy!
     redirect_to documents_path
